update and refactor postAffirmations thunk tests
